requirements.txt
aiogram==3.24.0
asyncpg==0.31.0
python-dotenv==1.2.1
yoomoney==0.1.2
apscheduler==3.11.2
requests==2.32.5